feat: 아이콘 토글 기능 추가
