+++ idearly/src/pages/AlgorithmSolvingPage/components/index.ts
@@ -1,1 +1,2 @@
-export * from './AlgorithmNav/AlgorithmNav'
+export * from "./AlgorithmNav/AlgorithmNav";
+export * from "./AlgorithmVoiceChat/AlgorithmVoiceChat";
